deps/debug_px4/pyproject.toml
[project]
name = "debug-px4"
version = "0.1.0"
description = "PX4 Zenoh debug subscriber with matplotlib visualization"
authors = [{ name = "Hanamy", email = "rongerch@outlook.com" }]
requires-python = ">=3.10"
dependencies = [
  "eclipse-zenoh>=1.0.0",
  "matplotlib",
  "mcap",
  "numpy",
]

[project.optional-dependencies]
dev = ["ruff"]

[project.scripts]
debug-px4 = "debug_px4.cli:main"
demo-fake-publisher = "demos.fake_publisher:main"
demo-mocap-logger = "demos.mocap_logger:main"
demo-mocap-reader = "demos.mocap_reader:main"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E402", "F841", "F401"]

[tool.hatch.build.targets.wheel]
packages = ["src/debug_px4", "demos"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
